refactor(daemon): move file accept into FileService, delete new_file_tx
Own-device file auto-accept was signalled to the daemon over an mpsc channel
(new_file_tx) drained by a spawn_file_auto_accept worker, because the FILES_ALPN
accept arm holds only field clones and couldn't reach MeshManager.

Move accept_file + try_auto_accept_file + is_own_device_sender onto FileService,
which now depends on Arc<Transport> (endpoint + blob store) and
Arc<NetworkRegistry> (the membership gate). The accept arm evaluates auto-accept
inline (it already runs in a per-connection task), so the channel + worker are
deleted. The IPC 'ray files accept' handler now delegates to self.files.

Build + clippy + 357 tests green.
